requirements.txt
requests
selenium
pyyaml
pytz
undetected_chromedriver 
seleniumbase